Refactor domHelper methods and import in index.js
