+++ src/index.js
@@ -1,17 +1,8 @@
 import './stylesheets/index.scss';
 import { v4 as uuidv4 } from 'uuid';
-import Todo from './models/todo';
 import renderProjectButton from './views/sidebar';
-import renderTodoItem from './views/todo';
 import renderProjectForm from './views/projectForm';
-
-const addTodoFormOpenBtn = document.getElementById('add-todo-form-open');
-const addTodoFormCloseBtn = document.getElementById('add-todo-form-close');
-const addTodoForm = document.getElementById('add-todo-form');
-const todoFormTitle = document.getElementById('add-todo-title');
-const todoFormDescription = document.getElementById('add-todo-description');
-const todoFormDueDate = document.getElementById('add-todo-date');
-const todoFormPriority = document.getElementById('add-todo-priority');
+import './utils/domHelpers';
 
 const projects = [
   {
@@ -80,33 +71,4 @@ window.addEventListener('DOMContentLoaded', () => {
   document
     .querySelector(`button[data-project-title="${activeProjectNow}"]`)
     .click();
-});
-
-// Create a new todo
-addTodoForm.addEventListener('submit', (event) => {
-  event.preventDefault();
-  const todo = new Todo(
-    todoFormTitle.value,
-    todoFormDescription.value,
-    todoFormDueDate.value,
-    todoFormPriority.value,
-  );
-
-  todo.addTodo();
-  const activeProject = JSON.parse(localStorage.getItem('activeProject'));
-  renderTodoItem(activeProject);
-  event.target.reset();
-  event.target.setAttribute('hidden', '');
-  addTodoFormOpenBtn.removeAttribute('hidden');
-});
-
-// Toggle todo form
-addTodoFormOpenBtn.addEventListener('click', (event) => {
-  addTodoForm.removeAttribute('hidden');
-  event.target.setAttribute('hidden', '');
-});
-
-addTodoFormCloseBtn.addEventListener('click', () => {
-  addTodoForm.setAttribute('hidden', '');
-  addTodoFormOpenBtn.removeAttribute('hidden');
 });
